pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sfrun"
authors = [
    {name = "Paresh Adhia", email = "padhia+github@gmail.com"},
]
description = "Snwoflake SQL runner"
readme = "README.md"
requires-python = ">=3.11"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "sfconn",
    "yappt",
    "snowflake-snowpark-python",
    "openpyxl"
]
dynamic = ["version"]

[project.scripts]
sfrun = "sfrun.main:cli"
sfrunb = "sfrun.batch:cli"

[tool.setuptools.dynamic]
version = {attr = "sfrun.util.__version__"}

[tool.setuptools]
packages = [
    "sfrun",
    "sfrun.formats"
]

[tool.ruff]
target-version = "py311"
line-length = 130
lint.extend-select = ["I"]

[tool.pytest.ini_options]
pythonpath = [ "." ]
filterwarnings = [ "ignore" ]

[tool.pyright]
reportUnknownMemberType = false
